removed models.py and split it into smaller files
