+++ ether_sql/models/__init.py
@@ -1,0 +1,4 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+
+base = declarative_base()
